Rearranged element positioning
Scrolling now takes place on the table. Header and footer views are now within
the scroll bars.
