+++ addon/styles/app.css
@@ -1,8 +1,9 @@
 .llama-table {
-	overflow: hidden;
+	overflow: auto;
 	width: 100%;
 	border-width: 1px;
 	border-style: solid;
+	position: relative;
 }
 
 .llama-columngroup {
@@ -41,6 +42,16 @@
 	border-style: solid;
 	border-width: 1px 0;
 	min-height: 30px;
+	position: absolute;
+	z-index: 2;
+}
+
+.llama-header {
+	top: 0;
+}
+
+.llama-footer {
+	bottom: 0;
 }
 
 .llama-header:first-child {
@@ -78,10 +89,14 @@
 }
 
 .llama-body {
-	overflow: auto;
 	min-height: 30px;
 	outline: none;
-	position: relative;
+}
+
+.llama-content {
+	padding-top: 31px;
+	padding-bottom: 31px;
+	z-index: 1;
 }
 
 .llama-empty {
